move tests into velo_action folder, remove proc_utils module.
