+++ velo_action/octopus/tests/test_octopus_client.py
@@ -3,7 +3,7 @@ import unittest.mock
 import pytest
 
 from velo_action.octopus.client import OctopusClient
-from velo_action.octopus.test_decorators import Request, mock_client_requests
+from velo_action.octopus.tests.test_decorators import Request, mock_client_requests
 
 
 @pytest.fixture
